create filter by price
