+++ db/seeds/02_products.sql
@@ -7,4 +7,8 @@ INSERT INTO products (seller_id, product_name, price, product_description, image
 INSERT INTO products (seller_id, product_name, price, product_description, image_url)
     VALUES (2, 'StrangeLove x Nike SB Dunk Low Pro QS "Valentine''s Day"', 1500, 'Deadstock. Trades accepted.', 'https://i.ytimg.com/vi/thPJ_0E0dPw/maxresdefault.jpg');
 
+INSERT INTO products (seller_id, product_name, price, product_description, image_url)
+    VALUES (2, 'Nike Dunk Low Retro', 300, 'Deadstock. Trades accepted.', 'https://sneakernews.com/wp-content/uploads/2021/01/Nike-Dunk-Low-Black-White-2021-Release-Info-0.jpg');
 
+INSERT INTO products (id, seller_id, product_name, price, product_description, image_url)
+    VALUES (4, 3, 'Jordan 1 Retro High OG', 50, 'Deadstock. Trades accepted.', 'https://blog.finishline.com/wp-content/uploads/2020/10/FNL_20201021_Jordan_Retro_AirJordan1HiOG_PhotoStudio_0066.jpg');
