+++ public/styles/filter.css
@@ -1,0 +1,3 @@
+img {
+  width: 500px;
+}
