+++ views/filters.ejs
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+<!--APP CSS-->
+<link rel="stylesheet" href="/styles/filter.css" type="text/css" />
+
+  <title>Document</title>
+</head>
+
+<%- include('partials/header') %>
+<body>
+
+<h1>Filtered Search</h1>
+
+<div class='filtered-products'>
+  <% for (let key of products) { %>
+    <a class="product" href="/products/<%= key.id %>"><img src="<%= key.image_url %>"></a>
+  <% } %>
+</div>
+
+</body>
+</html>
