+++ views/partials/header.ejs
@@ -1,9 +1,26 @@
+<!--External CSS-->
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" integrity="sha512-Fo3rlrZj/k7ujTnHg4CGR2D7kSs0v4LLanw2qksYuRlEzO+tcaEPQogQ0KaoGN26/zrn20ImR1DfuLWnOo7aBA==" crossorigin="anonymous" referrerpolicy="no-referrer" />
+
+<!--APP CSS-->
+<link rel="stylesheet" href="/styles/nav-header.css" type="text/css" />
+
 <header class="nav-header">
     <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
       <div class="navbar">
         <div class="left-side">
           <a class="nav-item" href="/">GraiLHLed</a>
-          <a class="nav-item" href="/filters">Filters</a>
+          <div class="search"></div>
+          <a class="nav-item" href="/filters">Max Price:</a>
+          <form method="POST" action="/filters">
+            <select id="prices" name="prices">
+              <option value="100">$100</option>
+              <option value="500">$500</option>
+              <option value="1000">$1000</option>
+              <option value="5000">$5000</option>
+              <option value="10000">ALL</option>
+            </select>
+            <button type="submit"><i class="fas fa-search"></i></button>
+          </form>
         </div>
         <div class="right-side">
           <a class="nav-item" href="/favourites">Favourites</a>
@@ -16,9 +33,4 @@
   </nav>
 
 </header>
-
-  <!--External CSS-->
-
-  <!--APP CSS-->
-  <link rel="stylesheet" href="/styles/nav-header.css" type="text/css" />
